+++ src/app/globals.css
@@ -21,6 +21,8 @@
 
       --secondary-darker: 210 40% 90.1%;
       --secondary-darker-foreground: 210 40% 10.1%;
+
+      --secondary-dark-transparent: rgba(210, 40%, 90.1%, 0.3);
       
       --muted: 210 40% 96.1%;
       --muted-foreground: 215.4 16.3% 46.9%;
@@ -65,6 +67,7 @@
 
       --secondary-darker : 217.2 32.6% 10.5%;
       --secondary-darker-foreground : 217.2 32.6% 15.5%;
+      --secondary-dark-transparent: rgba(217.2, 32.6%, 10.5%, 0.3);
 
       --muted: 217.2 32.6% 17.5%;
       --muted-foreground: 215 20.2% 65.1%;
