+++ src/app/groups/[groupId]/(pages)/history/columns.tsx
