+++ src/app/groups/[groupId]/(pages)/layout.tsx
@@ -8,19 +8,19 @@ export default function TabsLayout({ children }:{children: React.ReactNode}) {
   const { groupId } = useParams<{ groupId: string }>();
   const currentPath = usePathname();
 
-  const isActive = (path: string) => currentPath === path ? "bg-secondary rounded-lg" : ""; 
+  const isActive = (path: string) => currentPath === path ? "bg-primary rounded-full text-secondary" : ""; 
 
   return (
     <div className="flex flex-col h-[100dvh]">
     <div className="flex-grow">{children}</div>
-    <footer className="flex justify-around p-4 border-t mb-10">
+    <footer className="fixed bottom-0 left-0 right-0 flex justify-between items-center bg-secondarydark-transparent backdrop-blur-lg mb-6 rounded-full mx-6 drop-shadow-md">
       <Link href={`/groups/${groupId}/dashboard`} className={`${isActive(`/groups/${groupId}/dashboard`)} p-4`} >
         <Home/>
       </Link>
-      <Link href={`/groups/${groupId}/leaderboard`} className={`${isActive(`/groups/${groupId}/leaderboard`)}} p-4`}>
-        <span>👖</span>
+      <Link href={`/groups/${groupId}/leaderboard`} className={`flex items-center justify-center w-14 h-14 ${isActive(`/groups/${groupId}/leaderboard`)} p-4`}>
+        <span className="text-xl">👖</span>
       </Link>
-      <Link href={`/groups/${groupId}/history`} className={`${isActive(`/groups/${groupId}/hitory`)} p-4`}>
+      <Link href={`/groups/${groupId}/history`} className={`${isActive(`/groups/${groupId}/history`)} p-4`}>
         <History />
       </Link>
       <Link href={`/groups/${groupId}/stats`} className={`${isActive(`/groups/${groupId}/stats`)} p-4`}>
