+++ src/app/groups/[groupId]/(pages)/stats/page.tsx
